Add files via upload
This project focuses on practicing the automatic generation of C code from a Simulink model and implementing it on the ARM core of the ZYBO board. The specific task involves creating a Simulink model using block diagrams to calculate the nearest Fibonacci number to a given input. The Fibonacci series, defined as the sum of the two preceding numbers, serves as the basis for this model. The generated Simulink model will then be converted into C code using the Embedded Coder and tested on the ZYBO board.